requirements.txt
Markdown
accelerate==0.16.0
aiofiles
albumentations
basicsr
bitsandbytes==0.35.4
diffusers==0.12.1
discord-webhook
fastapi
ftfy
gitpython
jinja2
lion-pytorch>=0.0.4
matplotlib
pandas
modelcards
omegaconf
pytest
python-dotenv
python-multipart
requests
safetensors>=0.2.8
tensorboard
tensorflow-macos; sys_platform == 'darwin' and platform_machine == 'arm64'
tensorflow; sys_platform != 'darwin' or platform_machine != 'arm64'
tqdm
transformers>=4.26.1
uvicorn
websockets
xformers>=0.0.17.dev447